+++ src/game/blocks/!EXAMPLEBLOCK.ts
@@ -1,0 +1,10 @@
+// @ts-nocheck
+// ^^^ Remember to remove this!
+
+export class BLOCK extends BLOCKPARENT {
+    constructor(x: number, y: number, w: number, h: number) {
+        super(x, y, w, h);
+        this.texture = "INSERT_HERE";
+        this.type = "INSERT_HERE";
+    }
+}