+++ src/game/levels.ts
@@ -1,9 +1,10 @@
-const modules = import.meta.glob('/src/levels/**/*.json', { eager: true });
+const modules = import.meta.glob('/src/levels/**/*.json5', { eager: true });
 
 export const Levels = new Map<number, LevelData>();
 
 export interface LevelData {
     format: number,
+    type: string,
     about: LevelAbout,
     data: any[],
     spawn?: number[],
@@ -21,4 +22,10 @@ interface LevelAbout {
 for (const path in modules) {
     const level = modules[path] as LevelData;
     Levels.set(level.id, level);
+}
+
+export function GetLevel(lvlID: number) {
+    const lvl = Levels.get(lvlID);
+    if (!lvl) throw new Error("Could not get level with ID: " + lvlID);
+    return lvl;
 }