+++ src/main.ts
@@ -14,6 +14,7 @@ Input.addBind('right', ['KeyD', 'ArrowRight']);
 Input.addBind('left', ['KeyA', 'ArrowLeft']);
 Input.addBind('up', ['KeyW', 'ArrowUp']);
 Input.addBind('down', ['KeyS', 'ArrowDown']);
+Input.addBind('exit', ['KeyR']);
 
 // instead of master we testing webgl
 //GL.initialize(document.getElementById("game-canvas") as HTMLCanvasElement);