+++ vite.config.js
@@ -1,7 +1,8 @@
 import { defineConfig } from "vite"
 import { viteSingleFile } from "vite-plugin-singlefile"
+import json5Plugin from 'vite-plugin-json5';
 
 export default defineConfig({
-	plugins: [viteSingleFile()],
+	plugins: [viteSingleFile(), json5Plugin()],
 	base: '/GToH-4/',
 })